feat(devtools): tone down obtrusive elements in the Injector Tree tab
Some of the elements are a bit too large so they have been shrunken to fit in with the rest of the UI.
