+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/injector-tree/BUILD.bazel
@@ -1,7 +1,7 @@
-load("//devtools/tools:ng_module.bzl", "ng_module")
 load("@io_bazel_rules_sass//:defs.bzl", "sass_binary")
-load("//devtools/tools:typescript.bzl", "ts_library", "ts_test_library")
 load("//devtools/tools:defaults.bzl", "karma_web_test_suite")
+load("//devtools/tools:ng_module.bzl", "ng_module")
+load("//devtools/tools:typescript.bzl", "ts_library", "ts_test_library")
 
 package(default_visibility = ["//:__subpackages__"])
 
@@ -20,10 +20,10 @@ ng_module(
         "injector-tree.component.html",
     ],
     deps = [
-        ":injector_providers",
         ":injector_tree_fns",
         "//devtools/projects/ng-devtools/src/lib/devtools-tabs/dependency-injection:injector_tree_visualizer",
         "//devtools/projects/ng-devtools/src/lib/devtools-tabs/dependency-injection:resolution_path",
+        "//devtools/projects/ng-devtools/src/lib/devtools-tabs/injector-tree/injector-providers",
         "//devtools/projects/ng-devtools/src/lib/vendor/angular-split",
         "//devtools/projects/protocol",
         "//packages/common",
@@ -32,21 +32,6 @@ ng_module(
         "@npm//@types",
         "@npm//d3",
         "@npm//rxjs",
-    ],
-)
-
-ng_module(
-    name = "injector_providers",
-    srcs = [
-        "injector-providers.component.ts",
-    ],
-    deps = [
-        "//devtools/projects/ng-devtools/src/lib/devtools-tabs/dependency-injection:resolution_path",
-        "//devtools/projects/protocol",
-        "//packages/animations",
-        "//packages/common",
-        "//packages/core",
-        "@npm//@angular/material",
     ],
 )
