+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/injector-tree/injector-providers/injector-providers.component.html
@@ -1,0 +1,65 @@
+<h2 class="providers-title">Providers for {{ injector()?.name }}</h2>
+@if (injector()) {
+  <div class="injector-providers">
+    <div class="filter">
+      <mat-form-field appearance="outline">
+        <mat-label>Search by token</mat-label>
+        <input
+          type="text"
+          matInput
+          placeholder="Provider token"
+          (input)="searchToken.set($event.target.value)"
+          [value]="searchToken()"
+        />
+        <mat-icon matSuffix (click)="searchToken.set('')">close</mat-icon>
+      </mat-form-field>
+      <mat-form-field appearance="outline">
+        <mat-label>Search by type</mat-label>
+        <mat-select [value]="searchType()" (selectionChange)="searchType.set($event.value)">
+          <mat-option>None</mat-option>
+          @for (type of providerTypes; track type) {
+            <mat-option [value]="type">{{ $any(providerTypeToLabel)[type] }}</mat-option>
+          }
+        </mat-select>
+      </mat-form-field>
+    </div>
+    @if (visibleProviders().length > 0) {
+      <table mat-table [dataSource]="visibleProviders()" class="mat-elevation-z4">
+        <ng-container matColumnDef="token">
+          <th mat-header-cell *matHeaderCellDef><h3 class="column-title">Token</h3></th>
+          <td mat-cell *matCellDef="let provider">{{ provider.token }}</td>
+        </ng-container>
+        <ng-container matColumnDef="type">
+          <th mat-header-cell *matHeaderCellDef><h3 class="column-title">Type</h3></th>
+          <td mat-cell *matCellDef="let provider">
+            @if (provider.type === 'multi') {
+              multi (x{{ provider.index.length }})
+            } @else {
+              {{ $any(providerTypeToLabel)[provider.type] }}
+            }
+          </td>
+        </ng-container>
+        <ng-container matColumnDef="isViewProvider">
+          <th mat-header-cell *matHeaderCellDef><h3 class="column-title">Is View Provider</h3></th>
+          <td mat-cell *matCellDef="let provider">
+            <mat-icon>{{ provider.isViewProvider ? 'check_circle' : 'cancel' }}</mat-icon>
+          </td>
+        </ng-container>
+        <ng-container matColumnDef="log">
+          <th mat-header-cell *matHeaderCellDef><h3 class="column-title"></h3></th>
+          <td mat-cell *matCellDef="let provider">
+            <mat-icon
+              matTooltipPosition="left"
+              matTooltip="Log provider in console"
+              class="select"
+              (click)="select(provider)"
+              >send</mat-icon
+            >
+          </td>
+        </ng-container>
+        <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+        <tr mat-row *matRowDef="let row; columns: displayedColumns"></tr>
+      </table>
+    }
+  </div>
+}
