+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/injector-tree/injector-tree.component.scss
@@ -133,6 +133,30 @@ as-split-area {
   }
 }
 
+.options {
+  padding: 0.5rem;
+  display: flex;
+  align-items: center;
+  gap: 2rem;
+
+  mat-checkbox {
+    $checkbox-size: 14px;
+    --mdc-checkbox-state-layer-size: 2rem;
+    --mat-checkbox-label-text-size: 0.8rem;
+
+    ::ng-deep .mdc-checkbox {
+      width: $checkbox-size;
+      height: $checkbox-size;
+      flex: 0 0 $checkbox-size;
+
+      &__background {
+        width: $checkbox-size;
+        height: $checkbox-size;
+      }
+    }
+  }
+}
+
 .deps {
   display: flex;
   overflow: auto;
@@ -178,12 +202,22 @@ as-split-area {
   overflow: hidden;
 
   h2 {
-    padding: 4px 16px;
+    padding: 0.5rem 1rem;
     margin: 0;
-    height: 50px;
     display: flex;
     align-items: center;
     border-bottom: 1px solid #777777;
+    font-size: 0.875rem;
+    display: flex;
+    align-items: center;
+    gap: 0.375rem;
+
+    mat-icon {
+      font-size: 1rem;
+      width: 1rem;
+      height: 1rem;
+      display: block;
+    }
   }
 }
