+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/injector-tree/injector-tree.component.ts
@@ -33,7 +33,7 @@ import {
   InjectorTreeVisualizer,
 } from '../dependency-injection/injector-tree-visualizer';
 
-import {InjectorProvidersComponent} from './injector-providers.component';
+import {InjectorProvidersComponent} from './injector-providers/injector-providers.component';
 import {
   filterOutAngularInjectors,
   filterOutInjectorsWithNoProviders,
